small fixes

+++ database/migrations/2016_07_18_212854_create_work_orders_table.php
@@ -17,7 +17,7 @@ class CreateWorkOrdersTable extends Migration
             $table->unsignedInteger('invoice_id');
             $table->date('scheduled')->nullable();
             $table->text('description');
-            $table->unsignedTinyInteger('rate');
+            $table->unsignedInteger('rate');
             $table->boolean('completed');
             $table->timestamps();
